+++ R/mayoshinytheme.R
@@ -3,8 +3,11 @@
 #'
 #' Style a Shiny app with Mayo colors.
 #'
-#' @details Use this as the \code{theme = } argument to the main shiny page-builder functions, e.g., \code{\link[shiny]{fluidPage}}.
-#' @return Returns a file path reference to the Mayo Shiny CSS file, and attaches that path so shiny knows about it.
+#' @details Use this as the \code{theme = } argument to the main shiny
+#'     page-builder functions, e.g., \code{\link[shiny]{fluidPage}}.
+#' @return Returns a file path reference to the Mayo Shiny CSS file,
+#'     and attaches that path so shiny knows about it.
+#' @param theme Name of a theme.
 #' @seealso \code{\link[shinythemes]{shinytheme}}
 #' @examples
 #' \dontrun{
@@ -14,8 +17,16 @@
 #' )
 #' }
 #' @export
-mayoshinytheme <- function()
-{
+mayoshinytheme <- function(theme = c("classic", "horizon")) {
+
+  theme <- match.arg(theme)
+
+  if (theme == "horizon") {
+    file <- "bootstrap_horizon.css"
+  } else {
+    file <- "bootstrap.css"
+  }
+
   shiny::addResourcePath("mayoshinythemes", mayoshinypath())
-  file.path("mayoshinythemes", "www", "mayo_theme.css")
+  file.path("mayoshinythemes", "css", file)
 }