+++ inst/examples/setup.R
@@ -1,0 +1,8 @@
+
+#' Rows: 10
+
+# Create R Shiny template
+mayoshiny::use_mayoshiny("my_app")
+
+# View file structure
+fs::dir_tree("my_app")