+++ NEWS.md
@@ -1,4 +1,18 @@
-# mayoshiny v1.1.0
+# mayoshiny 2.0.0
+
+## Major changes
+
+* Use a custom bootstrap theme from `mayotheme`
+
+* Add a new `horizon` theme
+
+* Add a Mayo loader gif
+
+* `mayo_shiny` has been renamed to `use_mayoshiny`
+
+# mayoshiny 1.1.0
+
+## Minor changes
 
 * Added the NEWS file.